Reorganize tests structure
