+++ tests/plugins/index.js
